webrtc: close data channels cleanly
WebRTC data channel close is a synchronous close procedure. We close our
outgoing stream, in response the peer is expected to close its outgoing
stream. If the peer doesn't close its side of the stream we will end up
with a memory leak where the SCTP transport keeps reference to the
stream. So we check the number of invalid data channel closures and when
this goes over a threshold we close the connection.

For our custom purposes we can fork SCTP and implement a unilateral
stream Reset which is feasible because we anyway have a state machine
on top of the data channels. But for a RFC compliant SCTP
implementation, this is how the spec is supposed to work. SCTP stream
numbers are limited (uint16) so we do need to reuse the stream ids
forcing us to use a synchronous close mechanism.