Use previous well-known resource
Changing from "/.well-known/libp2p" to "/.well-known/libp2p/protocols" will break the ability to communicate with existing servers deployed without upgrading them first. Since these servers are running in places not under our control, they will not be immediately upgradable. This means that libp2phttp needs to support both the old and the new values, or revert to using the old value.

This PR reverts to using the old resource name, which was changed in #2757.